Move remaining error types, pull out new module for changelog representation types
