+++ src/Data/API/Changes.hs
@@ -55,9 +55,10 @@ module Data.API.Changes
     , prettyValueErrorPosition
     ) where
 
+import           Data.API.Changes.Types
+import           Data.API.Error
 import           Data.API.JSON
 import           Data.API.NormalForm
-import           Data.API.PP
 import           Data.API.Types
 import           Data.API.Utils
 import           Data.API.Value as Value
@@ -67,7 +68,6 @@ import           Control.Monad
 import qualified Data.Aeson as JS
 import qualified Data.ByteString.Char8 as B
 import qualified Data.ByteString.Base64 as B64
-import qualified Data.Graph as Graph
 import           Data.Map (Map)
 import qualified Data.Map as Map
 import           Data.Maybe
@@ -127,59 +127,7 @@ migrateDataDump' startApi endApi changelog custom root chks db = do
     db' <- applyChangesToDatabase' root custom' db changes  ?!? uncurry ValueError
     return (db', warnings)
 
-data MigrateFailure
-    = ValidateFailure ValidateFailure
-    | ValueError ValueError Position
-    deriving (Eq, Show)
 
-type MigrateWarning = ValidateWarning
-
-------------------
--- The key types
---
-
-type APIWithChangelog = (API, APIChangelog)
-
--- | An API changelog, consisting of a list of versions with the
--- changes from one version to the next.  The versions must be in
--- descending order (according to the 'Ord' 'Version' instance).
-data APIChangelog =
-     -- | The changes from the previous version up to this version.
-     ChangesUpTo VersionExtra [APIChange] APIChangelog
-     -- | The initial version
-   | ChangesStart Version
-    deriving (Eq, Show)
-
--- | A single change within a changelog
-data APIChange
-    = ChAddType       TypeName NormTypeDecl
-    | ChDeleteType    TypeName
-    | ChRenameType    TypeName TypeName
-
-      -- Specific changes for record types
-    | ChAddField      TypeName FieldName APIType (Maybe DefaultValue)
-    | ChDeleteField   TypeName FieldName
-    | ChRenameField   TypeName FieldName FieldName
-    | ChChangeField   TypeName FieldName APIType MigrationTag
-
-      -- Changes for union types
-    | ChAddUnionAlt    TypeName FieldName APIType
-    | ChDeleteUnionAlt TypeName FieldName
-    | ChRenameUnionAlt TypeName FieldName FieldName
-
-      -- Changes for enum types
-    | ChAddEnumVal    TypeName FieldName
-    | ChDeleteEnumVal TypeName FieldName
-    | ChRenameEnumVal TypeName FieldName FieldName
-
-      -- Custom migrations
-    | ChCustomType    TypeName MigrationTag
-    | ChCustomAll     MigrationTag
-    deriving (Eq, Show)
-
--- | Within the changelog, custom migrations are represented as
--- strings, so we have less type-safety.
-type MigrationTag = String
 
 -- | Custom migrations used in the changelog must be implemented in
 -- Haskell, and supplied in this record.  There are three kinds:
@@ -263,20 +211,6 @@ validateAfter chks _                  = chks >= CheckAll
 -- Changelog utils
 --
 
--- | Represents either a released version (with a version number) or
--- the version under development, which is newer than any release
-data VersionExtra = Release Version
-                  | DevVersion
-  deriving (Eq, Ord, Show)
-
-showVersionExtra :: VersionExtra -> String
-showVersionExtra (Release v) = showVersion v
-showVersionExtra DevVersion  = "development"
-
-instance PP VersionExtra where
-  pp = showVersionExtra
-
-
 -- | The earliest version in the changelog
 changelogStartVersion :: APIChangelog -> Version
 changelogStartVersion (ChangesStart v) = v
@@ -332,28 +266,6 @@ changeTags _                       = (Set.empty, Set.empty, Set.empty)
 -- Representing update positions
 --
 
--- | Represents the positions in a declaration to apply an update
-data UpdateDeclPos
-    = UpdateHere   (Maybe UpdateDeclPos)
-    | UpdateRecord (Map FieldName (Maybe UpdateTypePos))
-    | UpdateUnion  (Map FieldName (Maybe UpdateTypePos))
-    | UpdateType   UpdateTypePos
-    deriving (Eq, Show)
-
--- | Represents the positions in a type to apply an update
-data UpdateTypePos
-    = UpdateList UpdateTypePos
-    | UpdateMaybe UpdateTypePos
-    | UpdateNamed TypeName
-    deriving (Eq, Show)
-
-data APITableChange
-      -- | An initial API, an APIChange and the positions in which to apply it
-    = APIChange NormAPI APIChange (Map TypeName UpdateDeclPos)
-      -- | Request to validate the dataset against the given API
-    | ValidateData NormAPI
-    deriving (Eq, Show)
-
 -- | Given a type to be modified, find the positions in which each
 -- type in the API must be updated
 findUpdatePos :: TypeName -> NormAPI -> Map TypeName UpdateDeclPos
@@ -390,63 +302,6 @@ findUpdatePos tname api = Map.alter (Just . UpdateHere) tname $
 ---------------------------
 -- Validating API changes
 --
-
--- | Errors that may be discovered when validating a changelog
-data ValidateFailure
-        -- | the changelog must be in descending order of versions
-    = ChangelogOutOfOrder { vfLaterVersion   :: VersionExtra
-                          , vfEarlierVersion :: VersionExtra }
-        -- | forbid migrating from one version to an earlier version
-    | CannotDowngrade { vfFromVersion :: VersionExtra
-                      , vfToVersion   :: VersionExtra }
-        -- | an API uses types that are not declared
-    | ApiInvalid { vfInvalidVersion      :: VersionExtra
-                 , vfMissingDeclarations :: Set TypeName }
-        -- | changelog entry does not apply
-    | ChangelogEntryInvalid { vfSuccessfullyApplied :: [APITableChange]
-                            , vfFailedToApply       :: APIChange
-                            , vfApplyFailure        :: ApplyFailure }
-        -- | changelog is incomplete
-        --   (ie all entries apply ok but result isn't the target api)
-    | ChangelogIncomplete { vfChangelogVersion :: VersionExtra
-                          , vfTargetVersion    :: VersionExtra
-                          , vfDifferences      :: Map TypeName (MergeResult NormTypeDecl NormTypeDecl) }
-  deriving (Eq, Show)
-
-data ValidateWarning = ValidateWarning -- add warnings about bits we cannot check (opaque custom)
-  deriving Show
-
--- | Errors that may occur applying a single API change
-data ApplyFailure
-    = TypeExists           { afExistingType  :: TypeName }     -- ^ for adding or renaming type
-    | TypeDoesNotExist     { afMissingType   :: TypeName }     -- ^ for deleting or renaming a type
-    | TypeWrongKind        { afTypeName      :: TypeName
-                           , afExpectedKind  :: TypeKind }     -- ^ e.g. it's not a record type
-    | TypeInUse            { afTypeName      :: TypeName }     -- ^ cannot delete/modify types that are still used
-    | TypeMalformed        { afType          :: APIType
-                           , afMissingTypes  :: Set TypeName } -- ^ type refers to a non-existent type
-    | DeclMalformed        { afTypeName      :: TypeName
-                           , afDecl          :: NormTypeDecl
-                           , afMissingTypes  :: Set TypeName } -- ^ decl refers to a non-existent type
-    | FieldExists          { afTypeName      :: TypeName
-                           , afTypeKind      :: TypeKind
-                           , afExistingField :: FieldName }    -- ^ for adding or renaming a field
-    | FieldDoesNotExist    { afTypeName      :: TypeName
-                           , afTypeKind      :: TypeKind
-                           , afMissingField  :: FieldName }    -- ^ for deleting or renaming a field
-    | FieldBadDefaultValue { afTypeName      :: TypeName
-                           , afFieldName     :: FieldName
-                           , afFieldType     :: APIType
-                           , afBadDefault    :: DefaultValue } -- ^ for adding a field, must be a default
-                                                               --   value compatible with the type
-    | DefaultMissing       { afTypeName      :: TypeName
-                           , afFieldName     :: FieldName }    -- ^ for adding a field to a table
-    | TableChangeError     { afCustomMessage :: String }       -- ^ custom error in tableChange
-  deriving (Eq, Show)
-
-data TypeKind = TKRecord | TKUnion | TKEnum | TKNewtype | TKTypeSynonym
-  deriving (Eq, Show)
-
 
 -- | Check that a changelog adequately describes how to migrate from
 -- one version to another.
@@ -910,13 +765,6 @@ applyChangeToData' _ (ChDeleteEnumVal _ _)  _ v _ = pure v
 -- Utils for manipulating JS.Values
 --
 
--- | Errors that can be discovered when migrating data values
-data ValueError
-    = JSONError JSONError                  -- ^ Data doesn't match schema
-    | CustomMigrationError String JS.Value -- ^ Error generated during custom migration
-    | InvalidAPI ApplyFailure              -- ^ An API change was invalid
-    deriving (Eq, Show)
-
 withObject :: (JS.Object -> Position -> Either (ValueError, Position) JS.Object)
            -> JS.Value -> Position -> Either (ValueError, Position) JS.Value
 withObject alter (JS.Object obj) p = JS.Object <$> alter obj p
@@ -1115,197 +963,6 @@ matchesNormAPIDecl api d v0 p = case d of
     matchesNormAPIField ((fn, ty), (fn', v))
         | fn == fn' = matchesNormAPI api ty v (InField (_FieldName fn) : p)
         | otherwise = Left (JSONError (SyntaxError (unlines ["record out of order: ", show fn, show fn', show d, show v0])), p)
-
-
--------------------------------------
--- Utils for merging and diffing maps
---
-
-data MergeResult a b = OnlyInLeft a | InBoth a b | OnlyInRight b
-  deriving (Eq, Show)
-
-mergeMaps :: Ord k => Map k a -> Map k b -> Map k (MergeResult a b)
-mergeMaps m1 m2 = Map.unionWith (\(OnlyInLeft a) (OnlyInRight b) -> InBoth a b)
-                      (fmap OnlyInLeft m1) (fmap OnlyInRight m2)
-
-diffMaps :: (Eq a, Ord k) => Map k a -> Map k a -> Map k (MergeResult a a)
-diffMaps m1 m2 = Map.filter different $ mergeMaps m1 m2
-  where
-    different (InBoth a b) = a /= b
-    different _            = True
-
--- Attempts to match the keys of the maps to produce a map from keys
--- to pairs.
-matchMaps :: Ord k => Map k a -> Map k b -> Either (k, Either a b) (Map k (a, b))
-matchMaps m1 m2 = Map.traverseWithKey win $ mergeMaps m1 m2
-  where
-    win _ (InBoth x y)    = return (x, y)
-    win k (OnlyInLeft x)  = Left (k, Left x)
-    win k (OnlyInRight x) = Left (k, Right x)
-
-
--------------------------------------
--- Pretty-printing
---
-
-prettyMigrateFailure :: MigrateFailure -> String
-prettyMigrateFailure = unlines . ppLines
-
-prettyValidateFailure :: ValidateFailure -> String
-prettyValidateFailure = unlines . ppLines
-
-prettyValueError :: ValueError -> String
-prettyValueError = unlines . ppLines
-
-prettyValueErrorPosition :: (ValueError, Position) -> String
-prettyValueErrorPosition = unlines . ppLines
-
-
-instance PP TypeKind where
-  pp TKRecord      = "record"
-  pp TKUnion       = "union"
-  pp TKEnum        = "enum"
-  pp TKNewtype     = "newtype"
-  pp TKTypeSynonym = "type"
-
-ppATypeKind :: TypeKind -> String
-ppATypeKind TKRecord      = "a record"
-ppATypeKind TKUnion       = "a union"
-ppATypeKind TKEnum        = "an enum"
-ppATypeKind TKNewtype     = "a newtype"
-ppATypeKind TKTypeSynonym = "a type synonym"
-
-ppMemberWord :: TypeKind -> String
-ppMemberWord TKRecord      = "field"
-ppMemberWord TKUnion       = "alternative"
-ppMemberWord TKEnum        = "value"
-ppMemberWord TKNewtype     = "member"
-ppMemberWord TKTypeSynonym = "member"
-
-
-instance PPLines APIChange where
-  ppLines (ChAddType t d)           = ("added " ++ pp t ++ " ") `inFrontOf` ppLines d
-  ppLines (ChDeleteType t)          = ["removed " ++ pp t]
-  ppLines (ChRenameType t t')       = ["renamed " ++ pp t ++ " to " ++ pp t']
-  ppLines (ChAddField t f ty mb_v)  = [ "changed record " ++ pp t
-                                      , "  field added " ++ pp f ++ " :: " ++ pp ty
-                                        ++ maybe "" (\ v -> " default " ++ pp v) mb_v]
-  ppLines (ChDeleteField t f)       = ["changed record " ++ pp t, "  field removed " ++ pp f]
-  ppLines (ChRenameField t f f')    = [ "changed record " ++ pp t
-                                      , "  field renamed " ++ pp f ++ " to " ++ pp f']
-  ppLines (ChChangeField t f ty c)  = [ "changed record " ++ pp t
-                                      , "  field changed " ++ pp f ++ " :: " ++ pp ty
-                                        ++ " migration " ++ pp c]
-  ppLines (ChAddUnionAlt t f ty)    = [ "changed union " ++ pp t
-                                      , "  alternative added " ++ pp f ++ " :: " ++ pp ty]
-  ppLines (ChDeleteUnionAlt t f)    = [ "changed union " ++ pp t
-                                      , "  alternative removed " ++ pp f]
-  ppLines (ChRenameUnionAlt t f f') = [ "changed union " ++ pp t
-                                      , "  alternative renamed " ++ pp f ++ " to " ++ pp f']
-  ppLines (ChAddEnumVal t f)        = [ "changed enum " ++ pp t
-                                      , "  alternative added " ++ pp f]
-  ppLines (ChDeleteEnumVal t f)     = [ "changed enum " ++ pp t
-                                      , "  alternative removed " ++ pp f]
-  ppLines (ChRenameEnumVal t f f')  = [ "changed enum " ++ pp t
-                                      , "  alternative renamed " ++ pp f ++ " to " ++ pp f']
-  ppLines (ChCustomType t c)        = ["migration record " ++ pp t ++ " " ++ pp c]
-  ppLines (ChCustomAll c)           = ["migration " ++ pp c]
-
-instance PPLines MigrateFailure where
-  ppLines (ValidateFailure x) = ppLines x
-  ppLines (ValueError x ps)   = ppLines x ++ map prettyStep ps
-
-instance PPLines ValidateFailure where
-  ppLines (ChangelogOutOfOrder later earlier) =
-      ["Changelog out of order: version " ++ pp later
-           ++ " appears after version " ++ pp earlier]
-  ppLines (CannotDowngrade from to) =
-      ["Cannot downgrade from version " ++ pp from
-           ++ " to version " ++ pp to]
-  ppLines (ApiInvalid ver missing) =
-      ["Missing declarations in API version " ++ pp ver ++ ": " ++ pp missing]
-  ppLines (ChangelogEntryInvalid succs change af) =
-      ppLines af ++ ("when applying the change" : indent (ppLines change))
-          ++ if not (null succs)
-             then "after successfully applying the changes:"
-                  : indent (ppLines succs)
-             else []
-  ppLines (ChangelogIncomplete ver ver' diffs) =
-      ("Changelog incomplete! Differences between log version ("
-           ++ showVersionExtra ver ++ ") and latest version (" ++ showVersionExtra ver' ++ "):")
-      : indent (ppDiffs diffs)
-
-instance PPLines APITableChange where
-  ppLines (APIChange _ c _)  = ppLines c
-  ppLines (ValidateData _) = []
-
-ppDiffs :: Map TypeName (MergeResult NormTypeDecl NormTypeDecl) -> [String]
-ppDiffs  = concatMap (uncurry ppDiff) . sortDiffs . Map.toList
-
--- | Perform a topological sort of the differences, so that the
--- pretty-printed form can be copied directly into the changelog.
-sortDiffs :: [(TypeName, MergeResult NormTypeDecl NormTypeDecl)]
-          -> [(TypeName, MergeResult NormTypeDecl NormTypeDecl)]
-sortDiffs = reverse . Graph.flattenSCCs . Graph.stronglyConnComp . map f
-  where
-    f (tn, mr) = ((tn, mr), tn, Set.toList (mergeResultFreeVars mr))
-
-mergeResultFreeVars :: MergeResult NormTypeDecl NormTypeDecl -> Set TypeName
-mergeResultFreeVars (OnlyInLeft  x) = typeDeclFreeVars x
-mergeResultFreeVars (OnlyInRight x) = typeDeclFreeVars x
-mergeResultFreeVars (InBoth x y)    = typeDeclFreeVars x `Set.union` typeDeclFreeVars y
-
-ppDiff :: TypeName -> MergeResult NormTypeDecl NormTypeDecl -> [String]
-ppDiff t (OnlyInLeft _)  = ["removed " ++ pp t]
-ppDiff t (OnlyInRight d) = ("added " ++ pp t ++ " ") `inFrontOf` ppLines d
-ppDiff t (InBoth (NRecordType flds) (NRecordType flds')) =
-    ("changed record " ++ pp t)
-    : (concatMap (uncurry (ppDiffFields "field")) $ Map.toList $ diffMaps flds flds')
-ppDiff t (InBoth (NUnionType alts) (NUnionType alts')) =
-    ("changed union " ++ pp t)
-    : (concatMap (uncurry (ppDiffFields "alternative")) $ Map.toList $ diffMaps alts alts')
-ppDiff t (InBoth (NEnumType vals) (NEnumType vals')) =
-    ("changed enum " ++ pp t)
-    :  (map (\ x -> "  alternative removed " ++ pp x) $ Set.toList $ vals  Set.\\ vals')
-    ++ (map (\ x -> "  alternative added " ++ pp x)   $ Set.toList $ vals' Set.\\ vals)
-ppDiff t (InBoth _ _) = ["incompatible definitions of " ++ pp t]
-
-ppDiffFields :: String -> FieldName -> MergeResult APIType APIType -> [String]
-ppDiffFields s f (OnlyInLeft _)   = ["  " ++ s ++ " removed " ++ pp f]
-ppDiffFields s f (OnlyInRight ty) = ["  " ++ s ++ " added " ++ pp f ++ " :: " ++ pp ty]
-ppDiffFields s f (InBoth ty ty')   = [ "  incompatible types for " ++ s ++ " " ++ pp f
-                                     , "    changelog type:      " ++ pp ty
-                                     , "    latest version type: " ++ pp ty' ]
-
-
-instance PPLines ApplyFailure where
-  ppLines (TypeExists t)                  = ["Type " ++ pp t ++ " already exists"]
-  ppLines (TypeDoesNotExist t)            = ["Type " ++ pp t ++ " does not exist"]
-  ppLines (TypeWrongKind t k)             = ["Type " ++ pp t ++ " is not " ++ ppATypeKind k]
-  ppLines (TypeInUse t)                   = ["Type " ++ pp t ++ " is in use, so it cannot be modified"]
-  ppLines (TypeMalformed ty xs)           = ["Type " ++ pp ty
-                                             ++ " is malformed, missing declarations:"
-                                            , "  " ++ pp xs]
-  ppLines (DeclMalformed t _ xs)          = [ "Declaration of " ++ pp t
-                                              ++ " is malformed, missing declarations:"
-                                            , "  " ++ pp xs]
-  ppLines (FieldExists t k f)             = ["Type " ++ pp t ++ " already has the "
-                                             ++ ppMemberWord k ++ " " ++ pp f]
-  ppLines (FieldDoesNotExist t k f)       = ["Type " ++ pp t ++ " does not have the "
-                                             ++ ppMemberWord k ++ " " ++ pp f]
-  ppLines (FieldBadDefaultValue _ _ ty v) = ["Default value " ++ pp v
-                                             ++ " is not compatible with the type " ++ pp ty]
-  ppLines (DefaultMissing t f)            = ["Field " ++ pp f ++ " does not have a default value, but "
-                                             ++ pp t ++ " occurs in the database"]
-  ppLines (TableChangeError s)            = ["Error when detecting changed tables:", "  " ++ s]
-
-
-instance PPLines ValueError where
-  ppLines (JSONError e)              = [prettyJSONError e]
-  ppLines (CustomMigrationError e v) = [ "Custom migration error:", "  " ++ e
-                                       , "when migrating value"] ++ indent (ppLines v)
-  ppLines (InvalidAPI af)            = "Invalid API detected during value migration:"
-                                       : indent (ppLines af)
 
 
 -------------------------------------
